requirements.txt
Django==6.0
djangorestframework==3.16.1
channels==4.1.0
channels-redis==4.1.0
daphne==4.1.0
python-decouple==3.8
psycopg2-binary==2.9.11
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
requests==2.32.5
APScheduler==3.11.1
tzlocal==5.3.1
python-dateutil==2.8.2
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0
twilio==9.4.0
